requirements.txt
ray
torch
numpy<2
gymnasium==1.1.1
mujoco==2.3.7
imageio
PyYAML
matplotlib
